% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{addNumbers}
\alias{addNumbers}
\title{Add Numbers}
\usage{
addNumbers(arg1, arg2)
}
\arguments{
\item{arg1}{A number}

\item{arg2}{Another number}
}
\value{
A number
}
\description{
Add Numbers
}
\examples{
arg1 <- 5
arg2 <- 4
addNumbers(arg1, arg2)
}
